# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/webplus/getArticles": {
            "get": {
                "description": "按栏目、站点、时间分页获取文章",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "获取文章列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "栏目ID，逗号分隔",
                        "name": "columnId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "站点ID，逗号分隔",
                        "name": "siteId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "util.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        }
    }
}